partial refactor to add select instrument page
