--- src/components/Instrument.tsx
@@ -1,5 +1,0 @@
-import { TInstrumentProps } from "../types"
-
-export function Instrument({ name }: TInstrumentProps) {
-  return <div>{name}</div>
-}
